Renamed interface and moved factory methods
