+++ core/src/main/java/com/linecorp/armeria/server/ServiceConfigBuilder.java
@@ -44,7 +44,7 @@ import com.linecorp.armeria.common.annotation.Nullable;
 import com.linecorp.armeria.common.util.BlockingTaskExecutor;
 import com.linecorp.armeria.internal.common.websocket.WebSocketUtil;
 import com.linecorp.armeria.server.logging.AccessLogWriter;
-import com.linecorp.armeria.server.websocket.WebSocketService;
+import com.linecorp.armeria.server.websocket.DefaultWebSocketService;
 
 final class ServiceConfigBuilder implements ServiceConfigSetters {
 
@@ -321,7 +321,7 @@ final class ServiceConfigBuilder implements ServiceConfigSetters {
                                                                      unhandledExceptionsReporter);
         }
 
-        final boolean webSocket = service.as(WebSocketService.class) != null;
+        final boolean webSocket = service.as(DefaultWebSocketService.class) != null;
         final long requestTimeoutMillis;
         if (this.requestTimeoutMillis != null) {
             requestTimeoutMillis = this.requestTimeoutMillis;
